LayoutTests/platform/chromium/accessibility/loading-iframe-sends-notification-expected.txt

Before


After

End of test

This tests that when an iframe finishes loading, it sends a layout complete notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS subwebarea.childrenCount > 0 is false
Got notification on iframe: LayoutComplete
PASS newSubwebarea.childrenCount > 0 is true

TEST COMPLETE
